Split the Job.jsx into JobContent.jsx and JobCard.jsx
